started setting up modes! protagonist now has two modes: ranged and melee, with different behaviours. Need to change the way modes are changed asap (for now it's done by opening the inventory and clicking it)

+++ awsom_game/Entity.cpp
@@ -78,7 +78,7 @@ void Entity::CollideRect( RectF rect )
 	SetPos( hbx.TopLeft() );
 }
 
-void Entity::ApplyDamage( int dmg )
+void Entity::ApplyDamage( float dmg )
 {
 	if (!state.Is( State::Invincible ) && !state.Is( State::Damaged ) && IsAlive()) {
 		hp -= dmg;
